refactor: shared mnemonic components; multisig procedure explainers
- Extract MnemonicDisplay/MnemonicEntry into shared/components; reuse in
  create-wallet, import-wallet, and multisig wizard (restore mode now uses
  the word-grid entry instead of a textarea)
- Add procedure explainer boxes to wizard steps 1 and 4 clarifying that
  every co-signer runs the same wizard and swaps xpubs at the exchange step
- Reconcile msig_* keys across all 25 locales (drop msig_restore_words/
  msig_restore_invalid, reword msig_restore_info, add msig_procedure_info
  and msig_exchange_info)
