+++ web-wallet/src/assets/locales/bg.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Създай нова seed фраза за този мултисиг",
   "msig_seed_restore": "Вече имам seed фраза (възстановяване или повторно присъединяване)",
   "msig_backup_warning": "Запишете тези 24 думи и ги съхранявайте на сигурно място. Те са делът на този портфейл в мултисига — без тях (и публичните ключове на съподписващите) вашият дял не може да бъде възстановен.",
-  "msig_restore_info": "Въведете думите от seed фразата на вашия мултисиг дял, разделени с интервали.",
-  "msig_restore_words": "Seed думи (12 или 24)",
-  "msig_restore_invalid": "Това не е валидна seed фраза.",
+  "msig_restore_info": "Въведете думите от seed фразата на вашия мултисиг дял.",
+  "msig_procedure_info": "Всеки участник изпълнява същия този съветник на своята машина, всеки със собствена seed фраза. Ще спрете на стъпката за обмен на ключове, за да размените публичните си ключове със съподписващите — след това всички стигат до идентичния портфейл.",
+  "msig_exchange_info": "Вашите съподписващи виждат същата тази стъпка в своя съветник. Изпратете им вашия ключ и въведете техните по-долу — редът няма значение.",
   "msig_restore_no_verify": "Seed фразата е въведена ръчно — не е необходима проверка.",
   "msig_your_key": "Вашият ключ",
   "msig_your_key_info": "Споделете този публичен ключ с вашите съподписващи. Сам по себе си той не може да харчи нищо.",
