+++ web-wallet/src/assets/locales/ca.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Crear una llavor nova per a aquesta multisignatura",
   "msig_seed_restore": "Ja tinc una llavor (restaurar o reincorporar-se)",
   "msig_backup_warning": "Escriviu aquestes 24 paraules i guardeu-les de forma segura. Són la part d'aquesta cartera en la multisignatura — sense elles (i les claus públiques dels cosignataris) la vostra part no es pot recuperar.",
-  "msig_restore_info": "Introduïu les paraules de la llavor de la vostra part de la multisignatura, separades per espais.",
-  "msig_restore_words": "Paraules de la llavor (12 o 24)",
-  "msig_restore_invalid": "Aquesta no és una frase de llavor vàlida.",
+  "msig_restore_info": "Introduïu les paraules de la llavor de la vostra part de la multisignatura.",
+  "msig_procedure_info": "Cada participant executa aquest mateix assistent a la seva pròpia màquina, cadascun amb la seva pròpia llavor. Us aturareu al pas d'intercanvi de claus per intercanviar les claus públiques amb els vostres cosignataris — després tothom arriba a la mateixa cartera idèntica.",
+  "msig_exchange_info": "Els vostres cosignataris veuen aquest mateix pas al seu assistent. Envieu-los la vostra clau i introduïu les seves a sota — l'ordre no importa.",
   "msig_restore_no_verify": "Llavor introduïda manualment — no cal verificació.",
   "msig_your_key": "La vostra clau",
   "msig_your_key_info": "Compartiu aquesta clau pública amb els vostres cosignataris. Per si sola no pot gastar res.",
