+++ web-wallet/src/assets/locales/cs.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Vytvořit nový seed pro tento multisig",
   "msig_seed_restore": "Už mám seed (obnovení nebo opětovné připojení)",
   "msig_backup_warning": "Zapište si těchto 24 slov a uložte je na bezpečném místě. Jsou podílem této peněženky na multisigu — bez nich (a veřejných klíčů spolupodepisujících) nelze váš podíl obnovit.",
-  "msig_restore_info": "Zadejte slova seedu vašeho multisig podílu, oddělená mezerami.",
-  "msig_restore_words": "Slova seedu (12 nebo 24)",
-  "msig_restore_invalid": "Toto není platná seed fráze.",
+  "msig_restore_info": "Zadejte slova seedu vašeho multisig podílu.",
+  "msig_procedure_info": "Každý účastník spouští téhož průvodce na svém vlastním počítači, každý se svým vlastním seedem. Na kroku výměny klíčů se zastavíte, abyste si se svými spolupodepisujícími vyměnili veřejné klíče — poté všichni dospějí k identické peněžence.",
+  "msig_exchange_info": "Vaši spolupodepisující vidí tento stejný krok ve svém průvodci. Pošlete jim svůj klíč a jejich klíče zadejte níže — na pořadí nezáleží.",
   "msig_restore_no_verify": "Seed zadán ručně — ověření není potřeba.",
   "msig_your_key": "Váš klíč",
   "msig_your_key_info": "Sdílejte tento veřejný klíč se svými spolupodepisujícími. Sám o sobě nemůže nic utratit.",
