+++ web-wallet/src/assets/locales/de-de.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Einen neuen Seed fuer dieses Multisig erstellen",
   "msig_seed_restore": "Ich habe bereits einen Seed (wiederherstellen oder erneut beitreten)",
   "msig_backup_warning": "Schreibe diese 24 Woerter auf und bewahre sie sicher auf. Sie sind der Anteil dieses Wallets am Multisig — ohne sie (und die oeffentlichen Schluessel der Mitunterzeichner) kann dein Anteil nicht wiederhergestellt werden.",
-  "msig_restore_info": "Gib die Seed-Woerter deines Multisig-Anteils ein, getrennt durch Leerzeichen.",
-  "msig_restore_words": "Seed-Woerter (12 oder 24)",
-  "msig_restore_invalid": "Dies ist keine gueltige Seed-Phrase.",
+  "msig_restore_info": "Gib die Seed-Woerter deines Multisig-Anteils ein.",
+  "msig_procedure_info": "Jeder Teilnehmer durchlaeuft denselben Assistenten auf seinem eigenen Rechner, jeder mit seinem eigenen Seed. Beim Schritt zum Schluesselaustausch haeltst du an, um mit deinen Mitunterzeichnern die oeffentlichen Schluessel auszutauschen — danach gelangen alle zum identischen Wallet.",
+  "msig_exchange_info": "Deine Mitunterzeichner sehen denselben Schritt in ihrem Assistenten. Schicke ihnen deinen Schluessel und trage ihre unten ein — die Reihenfolge spielt keine Rolle.",
   "msig_restore_no_verify": "Seed manuell eingegeben — keine Ueberpruefung noetig.",
   "msig_your_key": "Dein Schluessel",
   "msig_your_key_info": "Teile diesen oeffentlichen Schluessel mit deinen Mitunterzeichnern. Er kann alleine nichts ausgeben.",
