+++ web-wallet/src/assets/locales/el.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Δημιουργία νέου seed για αυτό το multisig",
   "msig_seed_restore": "Έχω ήδη seed (ανάκτηση ή επανασύνδεση)",
   "msig_backup_warning": "Γράψτε αυτές τις 24 λέξεις και αποθηκεύστε τις με ασφάλεια. Είναι το μερίδιο αυτού του πορτοφολιού στο multisig — χωρίς αυτές (και τα δημόσια κλειδιά των συνυπογραφόντων) το μερίδιό σας δεν μπορεί να ανακτηθεί.",
-  "msig_restore_info": "Εισάγετε τις λέξεις seed του μεριδίου σας στο multisig, χωρισμένες με κενά.",
-  "msig_restore_words": "Λέξεις seed (12 ή 24)",
-  "msig_restore_invalid": "Αυτή δεν είναι έγκυρη φράση seed.",
+  "msig_restore_info": "Εισάγετε τις λέξεις seed του μεριδίου σας στο multisig.",
+  "msig_procedure_info": "Κάθε συμμετέχων εκτελεί τον ίδιο αυτόν οδηγό στο δικό του μηχάνημα, ο καθένας με το δικό του seed. Θα σταματήσετε στο βήμα ανταλλαγής κλειδιών για να ανταλλάξετε δημόσια κλειδιά με τους συνυπογράφοντές σας — έπειτα όλοι καταλήγουν στο ίδιο ακριβώς πορτοφόλι.",
+  "msig_exchange_info": "Οι συνυπογράφοντές σας βλέπουν το ίδιο αυτό βήμα στον οδηγό τους. Στείλτε τους το κλειδί σας και εισάγετε τα δικά τους παρακάτω — η σειρά δεν έχει σημασία.",
   "msig_restore_no_verify": "Το seed εισήχθη χειροκίνητα — δεν απαιτείται επαλήθευση.",
   "msig_your_key": "Το κλειδί σας",
   "msig_your_key_info": "Μοιραστείτε αυτό το δημόσιο κλειδί με τους συνυπογράφοντές σας. Από μόνο του δεν μπορεί να ξοδέψει τίποτα.",
