+++ web-wallet/src/assets/locales/en.json
@@ -1028,9 +1028,9 @@
   "msig_seed_new": "Create a new seed for this multisig",
   "msig_seed_restore": "I already have a seed (restore or rejoin)",
   "msig_backup_warning": "Write these 24 words down and store them safely. They are this wallet's share of the multisig — without them (and the co-signer public keys) your share cannot be recovered.",
-  "msig_restore_info": "Enter the seed words of your multisig share, separated by spaces.",
-  "msig_restore_words": "Seed words (12 or 24)",
-  "msig_restore_invalid": "This is not a valid seed phrase.",
+  "msig_restore_info": "Enter the seed words of your multisig share.",
+  "msig_procedure_info": "Every participant runs this same wizard on their own machine, each with their own seed. You will pause at the key-exchange step to swap public keys with your co-signers — then everyone arrives at the identical wallet.",
+  "msig_exchange_info": "Your co-signers see this same step in their wizard. Send them your key, and enter theirs below — the order does not matter.",
   "msig_restore_no_verify": "Seed entered manually — no verification needed.",
   "msig_your_key": "Your key",
   "msig_your_key_info": "Share this public key with your co-signers. It cannot spend anything on its own.",
