+++ web-wallet/src/assets/locales/es-es.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Crear una nueva semilla para esta multifirma",
   "msig_seed_restore": "Ya tengo una semilla (restaurar o reincorporarse)",
   "msig_backup_warning": "Escriba estas 24 palabras y guárdelas de forma segura. Son la parte de esta billetera en la multifirma — sin ellas (y las claves públicas de los cofirmantes) su parte no se puede recuperar.",
-  "msig_restore_info": "Ingrese las palabras de la semilla de su parte de la multifirma, separadas por espacios.",
-  "msig_restore_words": "Palabras de la semilla (12 o 24)",
-  "msig_restore_invalid": "Esta no es una frase semilla válida.",
+  "msig_restore_info": "Ingrese las palabras de la semilla de su parte de la multifirma.",
+  "msig_procedure_info": "Cada participante ejecuta este mismo asistente en su propia máquina, cada uno con su propia semilla. Se detendrá en el paso de intercambio de claves para intercambiar las claves públicas con sus cofirmantes — al final, todos llegan a una billetera idéntica.",
+  "msig_exchange_info": "Sus cofirmantes ven este mismo paso en su asistente. Envíeles su clave e ingrese las de ellos abajo — el orden no importa.",
   "msig_restore_no_verify": "Semilla ingresada manualmente — no se necesita verificación.",
   "msig_your_key": "Su clave",
   "msig_your_key_info": "Comparta esta clave pública con sus cofirmantes. Por sí sola no puede gastar nada.",
