+++ web-wallet/src/assets/locales/fi.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Luo uusi siemen tätä moniallekirjoitusta varten",
   "msig_seed_restore": "Minulla on jo siemen (palauta tai liity uudelleen)",
   "msig_backup_warning": "Kirjoita nämä 24 sanaa ylös ja säilytä ne turvallisesti. Ne ovat tämän lompakon osuus moniallekirjoituksesta — ilman niitä (ja yhteisallekirjoittajien julkisia avaimia) osuuttasi ei voi palauttaa.",
-  "msig_restore_info": "Syötä moniallekirjoitusosuutesi siemensanat välilyönnein eroteltuina.",
-  "msig_restore_words": "Siemensanat (12 tai 24)",
-  "msig_restore_invalid": "Tämä ei ole kelvollinen siemenlause.",
+  "msig_restore_info": "Syötä moniallekirjoitusosuutesi siemensanat.",
+  "msig_procedure_info": "Jokainen osallistuja suorittaa tämän saman ohjatun toiminnon omalla koneellaan, kukin omalla siemenellään. Pysähdyt avaintenvaihtovaiheeseen vaihtamaan julkiset avaimet yhteisallekirjoittajiesi kanssa — sen jälkeen kaikki päätyvät täsmälleen samaan lompakkoon.",
+  "msig_exchange_info": "Yhteisallekirjoittajasi näkevät tämän saman vaiheen omassa ohjatussa toiminnossaan. Lähetä heille avaimesi ja syötä heidän avaimensa alle — järjestyksellä ei ole väliä.",
   "msig_restore_no_verify": "Siemen syötetty käsin — vahvistusta ei tarvita.",
   "msig_your_key": "Oma avaimesi",
   "msig_your_key_info": "Jaa tämä julkinen avain yhteisallekirjoittajillesi. Se ei yksinään voi käyttää mitään varoja.",
