+++ web-wallet/src/assets/locales/fr.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Créer une nouvelle phrase de récupération pour ce multisig",
   "msig_seed_restore": "J'ai déjà une phrase de récupération (restaurer ou rejoindre)",
   "msig_backup_warning": "Écrivez ces 24 mots et conservez-les en sécurité. Ils constituent la part de ce portefeuille dans le multisig — sans eux (et sans les clés publiques des cosignataires), votre part ne peut pas être récupérée.",
-  "msig_restore_info": "Entrez les mots de la phrase de récupération de votre part du multisig, séparés par des espaces.",
-  "msig_restore_words": "Mots de la phrase de récupération (12 ou 24)",
-  "msig_restore_invalid": "Ce n'est pas une phrase de récupération valide.",
+  "msig_restore_info": "Entrez les mots de la phrase de récupération de votre part du multisig.",
+  "msig_procedure_info": "Chaque participant exécute ce même assistant sur sa propre machine, chacun avec sa propre phrase de récupération. Vous ferez une pause à l'étape d'échange des clés pour échanger les clés publiques avec vos cosignataires — ensuite, chacun obtient exactement le même portefeuille.",
+  "msig_exchange_info": "Vos cosignataires voient cette même étape dans leur assistant. Envoyez-leur votre clé et saisissez les leurs ci-dessous — l'ordre n'a pas d'importance.",
   "msig_restore_no_verify": "Phrase saisie manuellement — aucune vérification nécessaire.",
   "msig_your_key": "Votre clé",
   "msig_your_key_info": "Partagez cette clé publique avec vos cosignataires. Elle ne peut rien dépenser à elle seule.",
