+++ web-wallet/src/assets/locales/gl.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Crear unha nova semente para esta multisig",
   "msig_seed_restore": "Xa teño unha semente (restaurar ou reincorporarse)",
   "msig_backup_warning": "Escribe estas 24 palabras e gárdaas de forma segura. Son a parte desta carteira na multisig — sen elas (e as chaves públicas dos cofirmantes) a túa parte non se pode recuperar.",
-  "msig_restore_info": "Introduce as palabras da semente da túa parte da multisig, separadas por espazos.",
-  "msig_restore_words": "Palabras da semente (12 ou 24)",
-  "msig_restore_invalid": "Esta non é unha frase semente válida.",
+  "msig_restore_info": "Introduce as palabras da semente da túa parte da multisig.",
+  "msig_procedure_info": "Cada participante executa este mesmo asistente na súa propia máquina, cada un coa súa propia semente. Farás unha pausa no paso de intercambio de chaves para intercambiar as chaves públicas cos teus cofirmantes — despois, todos chegan á mesma carteira idéntica.",
+  "msig_exchange_info": "Os teus cofirmantes ven este mesmo paso no seu asistente. Envíalles a túa chave e introduce as deles abaixo — a orde non importa.",
   "msig_restore_no_verify": "Semente introducida manualmente — non se necesita verificación.",
   "msig_your_key": "A túa chave",
   "msig_your_key_info": "Comparte esta chave pública cos teus cofirmantes. Por si soa non pode gastar nada.",
