+++ web-wallet/src/assets/locales/hi.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "इस मल्टीसिग के लिए एक नई सीड बनाएं",
   "msig_seed_restore": "मेरे पास पहले से एक सीड है (पुनर्स्थापित करें या फिर से जुड़ें)",
   "msig_backup_warning": "इन 24 शब्दों को लिखें और सुरक्षित रूप से संग्रहीत करें। ये मल्टीसिग में इस वॉलेट का हिस्सा हैं — इनके बिना (और सह-हस्ताक्षरकर्ताओं की सार्वजनिक कुंजियों के बिना) आपका हिस्सा पुनर्प्राप्त नहीं किया जा सकता।",
-  "msig_restore_info": "अपने मल्टीसिग हिस्से की सीड के शब्द दर्ज करें, स्पेस से अलग करके।",
-  "msig_restore_words": "सीड शब्द (12 या 24)",
-  "msig_restore_invalid": "यह एक वैध सीड वाक्यांश नहीं है।",
+  "msig_restore_info": "अपने मल्टीसिग हिस्से की सीड के शब्द दर्ज करें।",
+  "msig_procedure_info": "प्रत्येक प्रतिभागी अपनी मशीन पर यही विज़ार्ड चलाता है, हर कोई अपनी-अपनी सीड के साथ। आप कुंजी आदान-प्रदान वाले चरण पर रुकेंगे ताकि अपने सह-हस्ताक्षरकर्ताओं के साथ सार्वजनिक कुंजियाँ बदल सकें — फिर सभी बिल्कुल एक जैसे वॉलेट पर पहुँचते हैं।",
+  "msig_exchange_info": "आपके सह-हस्ताक्षरकर्ता अपने विज़ार्ड में यही चरण देखते हैं। उन्हें अपनी कुंजी भेजें और नीचे उनकी कुंजियाँ दर्ज करें — क्रम मायने नहीं रखता।",
   "msig_restore_no_verify": "सीड मैन्युअल रूप से दर्ज की गई — सत्यापन की आवश्यकता नहीं।",
   "msig_your_key": "आपकी कुंजी",
   "msig_your_key_info": "इस सार्वजनिक कुंजी को अपने सह-हस्ताक्षरकर्ताओं के साथ साझा करें। यह अकेले कुछ भी खर्च नहीं कर सकती।",
