+++ web-wallet/src/assets/locales/hr.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Kreiraj novi seed za ovaj multisig",
   "msig_seed_restore": "Već imam seed (oporavak ili ponovno pridruživanje)",
   "msig_backup_warning": "Zapiši ove 24 riječi i čuvaj ih na sigurnom. One su udio ovog novčanika u multisigu — bez njih (i javnih ključeva supotpisnika) tvoj udio nije moguće oporaviti.",
-  "msig_restore_info": "Unesi seed riječi svog multisig udjela, odvojene razmacima.",
-  "msig_restore_words": "Seed riječi (12 ili 24)",
-  "msig_restore_invalid": "Ovo nije važeća seed fraza.",
+  "msig_restore_info": "Unesi seed riječi svog multisig udjela.",
+  "msig_procedure_info": "Svaki sudionik prolazi ovaj isti čarobnjak na svom računalu, svaki sa svojim seedom. Zastat ćeš na koraku razmjene ključeva da razmijeniš javne ključeve sa svojim supotpisnicima — zatim svi dolaze do identičnog novčanika.",
+  "msig_exchange_info": "Tvoji supotpisnici vide ovaj isti korak u svom čarobnjaku. Pošalji im svoj ključ i unesi njihove ključeve ispod — redoslijed nije bitan.",
   "msig_restore_no_verify": "Seed je unesen ručno — provjera nije potrebna.",
   "msig_your_key": "Tvoj ključ",
   "msig_your_key_info": "Podijeli ovaj javni ključ sa svojim supotpisnicima. Sam po sebi ne može ništa trošiti.",
