+++ web-wallet/src/assets/locales/id.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Buat seed baru untuk multisig ini",
   "msig_seed_restore": "Saya sudah punya seed (pulihkan atau bergabung kembali)",
   "msig_backup_warning": "Tulis 24 kata ini dan simpan dengan aman. Kata-kata ini adalah bagian dompet ini dalam multisig — tanpanya (dan kunci publik rekan penandatangan) bagian Anda tidak dapat dipulihkan.",
-  "msig_restore_info": "Masukkan kata seed dari bagian multisig Anda, dipisahkan dengan spasi.",
-  "msig_restore_words": "Kata seed (12 atau 24)",
-  "msig_restore_invalid": "Ini bukan frasa seed yang valid.",
+  "msig_restore_info": "Masukkan kata seed dari bagian multisig Anda.",
+  "msig_procedure_info": "Setiap peserta menjalankan wizard yang sama ini di mesin mereka sendiri, masing-masing dengan seed mereka sendiri. Anda akan berhenti sejenak di langkah tukar kunci untuk saling bertukar kunci publik dengan rekan penandatangan Anda — kemudian semua orang sampai pada dompet yang identik.",
+  "msig_exchange_info": "Rekan penandatangan Anda melihat langkah yang sama ini di wizard mereka. Kirimkan kunci Anda kepada mereka, dan masukkan kunci mereka di bawah — urutannya tidak penting.",
   "msig_restore_no_verify": "Seed dimasukkan secara manual — tidak perlu verifikasi.",
   "msig_your_key": "Kunci Anda",
   "msig_your_key_info": "Bagikan kunci publik ini kepada rekan penandatangan Anda. Kunci ini tidak dapat membelanjakan apa pun sendirian.",
