+++ web-wallet/src/assets/locales/it.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Crea un nuovo seed per questo multisig",
   "msig_seed_restore": "Ho già un seed (ripristino o riadesione)",
   "msig_backup_warning": "Scrivi queste 24 parole e conservale in modo sicuro. Sono la quota di questo wallet nel multisig — senza di esse (e le chiavi pubbliche dei cofirmatari) la tua quota non può essere recuperata.",
-  "msig_restore_info": "Inserisci le parole seed della tua quota multisig, separate da spazi.",
-  "msig_restore_words": "Parole seed (12 o 24)",
-  "msig_restore_invalid": "Questa non è una frase seed valida.",
+  "msig_restore_info": "Inserisci le parole seed della tua quota multisig.",
+  "msig_procedure_info": "Ogni partecipante esegue questa stessa procedura guidata sulla propria macchina, ciascuno con il proprio seed. Ti fermerai al passaggio dello scambio delle chiavi per scambiare le chiavi pubbliche con i tuoi cofirmatari — alla fine tutti ottengono lo stesso identico wallet.",
+  "msig_exchange_info": "I tuoi cofirmatari vedono questo stesso passaggio nella loro procedura guidata. Invia loro la tua chiave e inserisci le loro qui sotto — l'ordine non ha importanza.",
   "msig_restore_no_verify": "Seed inserito manualmente — nessuna verifica necessaria.",
   "msig_your_key": "La tua chiave",
   "msig_your_key_info": "Condividi questa chiave pubblica con i tuoi cofirmatari. Da sola non può spendere nulla.",
