+++ web-wallet/src/assets/locales/ja.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "このマルチシグ用に新しいシードを作成する",
   "msig_seed_restore": "すでにシードを持っています（復元または再参加）",
   "msig_backup_warning": "これらの24単語を書き留めて、安全に保管してください。これはこのウォレットのマルチシグにおける持ち分です — これら（および共同署名者の公開鍵）がないと、あなたの持ち分は復元できません。",
-  "msig_restore_info": "マルチシグの持ち分のシード単語を、スペースで区切って入力してください。",
-  "msig_restore_words": "シード単語（12 または 24）",
-  "msig_restore_invalid": "有効なシードフレーズではありません。",
+  "msig_restore_info": "マルチシグの持ち分のシード単語を入力してください。",
+  "msig_procedure_info": "各参加者は自分のマシンで、それぞれ自分のシードを使ってこの同じウィザードを実行します。鍵の交換ステップで一時停止し、共同署名者と公開鍵を交換します — その後、全員が同一のウォレットにたどり着きます。",
+  "msig_exchange_info": "共同署名者も自分のウィザードでこれと同じステップを見ています。あなたの鍵を送り、下に相手の鍵を入力してください — 順序は関係ありません。",
   "msig_restore_no_verify": "シードは手動で入力されました — 確認は不要です。",
   "msig_your_key": "あなたの鍵",
   "msig_your_key_info": "この公開鍵を共同署名者と共有してください。この鍵単独では何も支出できません。",
