+++ web-wallet/src/assets/locales/lt.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Sukurti naują seed šiam multisig",
   "msig_seed_restore": "Jau turiu seed (atkurti arba prisijungti iš naujo)",
   "msig_backup_warning": "Užsirašykite šiuos 24 žodžius ir saugokite juos saugiai. Jie yra šios piniginės multisig dalis — be jų (ir bendrapasirašytojų viešųjų raktų) jūsų dalies atkurti neįmanoma.",
-  "msig_restore_info": "Įveskite savo multisig dalies seed žodžius, atskirtus tarpais.",
-  "msig_restore_words": "Seed žodžiai (12 arba 24)",
-  "msig_restore_invalid": "Tai nėra galiojanti seed frazė.",
+  "msig_restore_info": "Įveskite savo multisig dalies seed žodžius.",
+  "msig_procedure_info": "Kiekvienas dalyvis vykdo šį patį vedlį savo kompiuteryje, kiekvienas su savo seed. Raktų apsikeitimo žingsnyje stabtelėsite, kad apsikeistumėte viešaisiais raktais su bendrapasirašytojais — tada visi gauna identišką piniginę.",
+  "msig_exchange_info": "Jūsų bendrapasirašytojai savo vedlyje mato tą patį žingsnį. Nusiųskite jiems savo raktą ir žemiau įveskite jų raktus — tvarka nesvarbi.",
   "msig_restore_no_verify": "Seed įvestas rankiniu būdu — patikrinimas nereikalingas.",
   "msig_your_key": "Jūsų raktas",
   "msig_your_key_info": "Pasidalinkite šiuo viešuoju raktu su savo bendrapasirašytojais. Pats vienas jis negali nieko išleisti.",
