+++ web-wallet/src/assets/locales/nl.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Een nieuwe seed voor deze multisig aanmaken",
   "msig_seed_restore": "Ik heb al een seed (herstellen of opnieuw deelnemen)",
   "msig_backup_warning": "Schrijf deze 24 woorden op en bewaar ze veilig. Ze zijn het aandeel van deze portemonnee in de multisig — zonder deze woorden (en de publieke sleutels van de mede-ondertekenaars) kan uw aandeel niet worden hersteld.",
-  "msig_restore_info": "Voer de seedwoorden van uw multisig-aandeel in, gescheiden door spaties.",
-  "msig_restore_words": "Seedwoorden (12 of 24)",
-  "msig_restore_invalid": "Dit is geen geldige seed phrase.",
+  "msig_restore_info": "Voer de seedwoorden van uw multisig-aandeel in.",
+  "msig_procedure_info": "Elke deelnemer doorloopt dezelfde wizard op zijn eigen computer, ieder met een eigen seed. Bij de stap voor het uitwisselen van sleutels pauzeert u om publieke sleutels met uw mede-ondertekenaars uit te wisselen — daarna komt iedereen uit bij dezelfde portemonnee.",
+  "msig_exchange_info": "Uw mede-ondertekenaars zien dezelfde stap in hun wizard. Stuur hun uw sleutel en voer hieronder de hunne in — de volgorde maakt niet uit.",
   "msig_restore_no_verify": "Seed handmatig ingevoerd — geen verificatie nodig.",
   "msig_your_key": "Uw sleutel",
   "msig_your_key_info": "Deel deze publieke sleutel met uw mede-ondertekenaars. Hij kan op zichzelf niets uitgeven.",
