+++ web-wallet/src/assets/locales/pl.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Utwórz nowy seed dla tego multisig",
   "msig_seed_restore": "Mam już seed (odzyskiwanie lub ponowne dołączenie)",
   "msig_backup_warning": "Zapisz te 24 słowa i przechowuj je bezpiecznie. Stanowią one udział tego portfela w multisig — bez nich (oraz kluczy publicznych współsygnatariuszy) Twojego udziału nie da się odzyskać.",
-  "msig_restore_info": "Wprowadź słowa seed swojego udziału w multisig, oddzielone spacjami.",
-  "msig_restore_words": "Słowa seed (12 lub 24)",
-  "msig_restore_invalid": "To nie jest prawidłowa fraza seed.",
+  "msig_restore_info": "Wprowadź słowa seed swojego udziału w multisig.",
+  "msig_procedure_info": "Każdy uczestnik przechodzi ten sam kreator na własnym komputerze, każdy z własnym seedem. Zatrzymasz się na etapie wymiany kluczy, aby wymienić klucze publiczne ze współsygnatariuszami — a potem wszyscy dojdą do identycznego portfela.",
+  "msig_exchange_info": "Twoi współsygnatariusze widzą ten sam krok w swoim kreatorze. Wyślij im swój klucz i wprowadź poniżej ich klucze — kolejność nie ma znaczenia.",
   "msig_restore_no_verify": "Seed wprowadzony ręcznie — weryfikacja nie jest potrzebna.",
   "msig_your_key": "Twój klucz",
   "msig_your_key_info": "Udostępnij ten klucz publiczny swoim współsygnatariuszom. Sam w sobie nie pozwala niczego wydać.",
