+++ web-wallet/src/assets/locales/pt-br.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Criar uma nova seed para este multisig",
   "msig_seed_restore": "Já tenho uma seed (restaurar ou reingressar)",
   "msig_backup_warning": "Escreva estas 24 palavras e guarde-as com segurança. Elas são a parte desta carteira no multisig — sem elas (e as chaves públicas dos cossignatários), sua parte não poderá ser recuperada.",
-  "msig_restore_info": "Digite as palavras da seed da sua parte do multisig, separadas por espaços.",
-  "msig_restore_words": "Palavras da seed (12 ou 24)",
-  "msig_restore_invalid": "Esta não é uma frase seed válida.",
+  "msig_restore_info": "Digite as palavras da seed da sua parte do multisig.",
+  "msig_procedure_info": "Cada participante executa este mesmo assistente em sua própria máquina, cada um com sua própria seed. Você fará uma pausa na etapa de troca de chaves para trocar chaves públicas com seus cossignatários — então todos chegam a uma carteira idêntica.",
+  "msig_exchange_info": "Seus cossignatários veem esta mesma etapa no assistente deles. Envie a sua chave para eles e insira as deles abaixo — a ordem não importa.",
   "msig_restore_no_verify": "Seed digitada manualmente — nenhuma verificação necessária.",
   "msig_your_key": "Sua chave",
   "msig_your_key_info": "Compartilhe esta chave pública com seus cossignatários. Sozinha, ela não pode gastar nada.",
