+++ web-wallet/src/assets/locales/ro.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Creează un seed nou pentru acest multisig",
   "msig_seed_restore": "Am deja un seed (recuperare sau realăturare)",
   "msig_backup_warning": "Scrieți aceste 24 de cuvinte și păstrați-le în siguranță. Ele sunt partea acestui portofel din multisig — fără ele (și fără cheile publice ale co-semnatarilor), partea dumneavoastră nu poate fi recuperată.",
-  "msig_restore_info": "Introduceți cuvintele seed ale părții dumneavoastră din multisig, separate prin spații.",
-  "msig_restore_words": "Cuvinte seed (12 sau 24)",
-  "msig_restore_invalid": "Aceasta nu este o frază seed validă.",
+  "msig_restore_info": "Introduceți cuvintele seed ale părții dumneavoastră din multisig.",
+  "msig_procedure_info": "Fiecare participant parcurge același asistent pe propriul calculator, fiecare cu propriul seed. Vă veți opri la pasul de schimb de chei pentru a face schimb de chei publice cu co-semnatarii — apoi toți ajung la un portofel identic.",
+  "msig_exchange_info": "Co-semnatarii dumneavoastră văd același pas în asistentul lor. Trimiteți-le cheia dumneavoastră și introduceți-le pe ale lor mai jos — ordinea nu contează.",
   "msig_restore_no_verify": "Seed introdus manual — nu este necesară verificarea.",
   "msig_your_key": "Cheia dumneavoastră",
   "msig_your_key_info": "Partajați această cheie publică cu co-semnatarii dumneavoastră. Singură, nu poate cheltui nimic.",
