+++ web-wallet/src/assets/locales/ru.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Создать новый seed для этой мультиподписи",
   "msig_seed_restore": "У меня уже есть seed (восстановление или повторное присоединение)",
   "msig_backup_warning": "Запишите эти 24 слова и храните их в безопасном месте. Они являются долей этого кошелька в мультиподписи — без них (и публичных ключей соподписантов) вашу долю восстановить невозможно.",
-  "msig_restore_info": "Введите слова seed-фразы вашей доли мультиподписи, разделяя их пробелами.",
-  "msig_restore_words": "Слова seed-фразы (12 или 24)",
-  "msig_restore_invalid": "Это не действительная seed-фраза.",
+  "msig_restore_info": "Введите слова seed-фразы вашей доли мультиподписи.",
+  "msig_procedure_info": "Каждый участник проходит этот же мастер на своём компьютере, каждый со своим seed. На шаге обмена ключами вы остановитесь, чтобы обменяться публичными ключами с соподписантами, — после этого у всех получится идентичный кошелёк.",
+  "msig_exchange_info": "Ваши соподписанты видят этот же шаг в своём мастере. Отправьте им свой ключ и введите их ключи ниже — порядок не имеет значения.",
   "msig_restore_no_verify": "Seed введён вручную — проверка не требуется.",
   "msig_your_key": "Ваш ключ",
   "msig_your_key_info": "Передайте этот публичный ключ вашим соподписантам. Сам по себе он не позволяет ничего потратить.",
