+++ web-wallet/src/assets/locales/sk.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Vytvoriť nový seed pre tento multisig",
   "msig_seed_restore": "Už mám seed (obnovenie alebo opätovné pripojenie)",
   "msig_backup_warning": "Zapíšte si týchto 24 slov a uchovajte ich na bezpečnom mieste. Sú podielom tejto peňaženky na multisigu — bez nich (a verejných kľúčov spolupodpisovateľov) sa váš podiel nedá obnoviť.",
-  "msig_restore_info": "Zadajte slová seedu vášho multisig podielu, oddelené medzerami.",
-  "msig_restore_words": "Slová seedu (12 alebo 24)",
-  "msig_restore_invalid": "Toto nie je platná seed fráza.",
+  "msig_restore_info": "Zadajte slová seedu vášho multisig podielu.",
+  "msig_procedure_info": "Každý účastník prechádza tým istým sprievodcom na vlastnom počítači, každý s vlastným seedom. Pri kroku výmeny kľúčov sa zastavíte, aby ste si so spolupodpisovateľmi vymenili verejné kľúče — potom všetci dospejú k identickej peňaženke.",
+  "msig_exchange_info": "Vaši spolupodpisovatelia vidia ten istý krok vo svojom sprievodcovi. Pošlite im svoj kľúč a ich kľúče zadajte nižšie — na poradí nezáleží.",
   "msig_restore_no_verify": "Seed zadaný ručne — overenie nie je potrebné.",
   "msig_your_key": "Váš kľúč",
   "msig_your_key_info": "Zdieľajte tento verejný kľúč so svojimi spolupodpisovateľmi. Sám osebe nemôže nič minúť.",
