+++ web-wallet/src/assets/locales/sr.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Креирај нови seed за овај multisig",
   "msig_seed_restore": "Већ имам seed (опоравак или поновно придруживање)",
   "msig_backup_warning": "Запишите ове 24 речи и чувајте их на сигурном месту. Оне су удео овог новчаника у multisig-у — без њих (и јавних кључева супотписника) ваш удео не може бити опорављен.",
-  "msig_restore_info": "Унесите seed речи вашег multisig удела, раздвојене размацима.",
-  "msig_restore_words": "Seed речи (12 или 24)",
-  "msig_restore_invalid": "Ово није важећа seed фраза.",
+  "msig_restore_info": "Унесите seed речи вашег multisig удела.",
+  "msig_procedure_info": "Сваки учесник пролази кроз овај исти чаробњак на свом рачунару, свако са својим seed-ом. Застаћете на кораку размене кључева да бисте разменили јавне кључеве са својим супотписницима — након тога сви долазе до идентичног новчаника.",
+  "msig_exchange_info": "Ваши супотписници виде овај исти корак у свом чаробњаку. Пошаљите им свој кључ и унесите њихове кључеве испод — редослед није битан.",
   "msig_restore_no_verify": "Seed је унет ручно — провера није потребна.",
   "msig_your_key": "Ваш кључ",
   "msig_your_key_info": "Поделите овај јавни кључ са својим супотписницима. Сам по себи не може ништа трошити.",
