+++ web-wallet/src/assets/locales/tr.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Bu çoklu imza için yeni bir seed oluştur",
   "msig_seed_restore": "Zaten bir seed'im var (kurtar veya yeniden katıl)",
   "msig_backup_warning": "Bu 24 kelimeyi yazın ve güvenli bir yerde saklayın. Bunlar bu cüzdanın çoklu imzadaki payıdır — onlar (ve ortak imzalayanların açık anahtarları) olmadan payınız kurtarılamaz.",
-  "msig_restore_info": "Çoklu imza payınızın seed kelimelerini boşluklarla ayırarak girin.",
-  "msig_restore_words": "Seed kelimeleri (12 veya 24)",
-  "msig_restore_invalid": "Bu geçerli bir seed ifadesi değil.",
+  "msig_restore_info": "Çoklu imza payınızın seed kelimelerini girin.",
+  "msig_procedure_info": "Her katılımcı bu aynı sihirbazı kendi makinesinde, kendi seed'iyle çalıştırır. Anahtar değişimi adımında durarak ortak imzalayanlarınızla açık anahtarları takas edeceksiniz — ardından herkes birebir aynı cüzdana ulaşır.",
+  "msig_exchange_info": "Ortak imzalayanlarınız kendi sihirbazlarında bu aynı adımı görür. Anahtarınızı onlara gönderin ve onlarınkini aşağıya girin — sıralamanın önemi yoktur.",
   "msig_restore_no_verify": "Seed elle girildi — doğrulama gerekmiyor.",
   "msig_your_key": "Anahtarınız",
   "msig_your_key_info": "Bu açık anahtarı ortak imzalayanlarınızla paylaşın. Tek başına hiçbir şey harcayamaz.",
