+++ web-wallet/src/assets/locales/uk.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "Створити новий seed для цього мультипідпису",
   "msig_seed_restore": "У мене вже є seed (відновлення або повторне приєднання)",
   "msig_backup_warning": "Запишіть ці 24 слова та зберігайте їх у безпечному місці. Вони є часткою цього гаманця в мультипідписі — без них (і публічних ключів співпідписантів) вашу частку неможливо відновити.",
-  "msig_restore_info": "Введіть слова seed вашої частки мультипідпису, розділені пробілами.",
-  "msig_restore_words": "Слова seed (12 або 24)",
-  "msig_restore_invalid": "Це недійсна seed-фраза.",
+  "msig_restore_info": "Введіть слова seed вашої частки мультипідпису.",
+  "msig_procedure_info": "Кожен учасник проходить цей самий майстер на своєму комп'ютері, кожен зі своїм seed. На кроці обміну ключами ви зупинитеся, щоб обмінятися публічними ключами зі співпідписантами, — після цього всі отримають ідентичний гаманець.",
+  "msig_exchange_info": "Ваші співпідписанти бачать цей самий крок у своєму майстрі. Надішліть їм свій ключ і введіть їхні ключі нижче — порядок не має значення.",
   "msig_restore_no_verify": "Seed введено вручну — перевірка не потрібна.",
   "msig_your_key": "Ваш ключ",
   "msig_your_key_info": "Поділіться цим публічним ключем зі своїми співпідписантами. Сам по собі він не може нічого витратити.",
