+++ web-wallet/src/assets/locales/zh-cn.json
@@ -1027,9 +1027,9 @@
   "msig_seed_new": "为此多重签名创建新种子",
   "msig_seed_restore": "我已有种子（恢复或重新加入）",
   "msig_backup_warning": "写下这24个单词并妥善保管。它们是此钱包在多重签名中的份额 — 没有它们（以及共同签名者的公钥），您的份额将无法恢复。",
-  "msig_restore_info": "输入您的多重签名份额的种子单词，以空格分隔。",
-  "msig_restore_words": "种子单词（12 或 24 个）",
-  "msig_restore_invalid": "这不是有效的种子短语。",
+  "msig_restore_info": "输入您的多重签名份额的种子单词。",
+  "msig_procedure_info": "每位参与者都在自己的设备上运行这个相同的向导，各自使用自己的种子。您将在交换密钥步骤暂停，与共同签名者互换公钥 — 之后所有人将得到完全相同的钱包。",
+  "msig_exchange_info": "您的共同签名者在其向导中也会看到这一步骤。将您的密钥发送给他们，并在下方输入他们的密钥 — 顺序无关紧要。",
   "msig_restore_no_verify": "种子为手动输入 — 无需验证。",
   "msig_your_key": "您的密钥",
   "msig_your_key_info": "将此公钥分享给您的共同签名者。它本身无法花费任何资金。",
